fixed state when multiple session cookies are received

+++ pygrim/components/session/file_session_storage.py
@@ -24,7 +24,7 @@ class FileSessionStorage(SessionStorage):
 
     def delete(self, session):
         full_path = self._get_path(session.get_id())
-        if os.path.isfile(full_path):
+        if full_path and os.path.isfile(full_path):
             try:
                 os.remove(full_path)
                 successful = True
@@ -45,8 +45,9 @@ class FileSessionStorage(SessionStorage):
     def load(self, request):
         session_id, session_new = self._get_id(request)
         try:
-            if os.path.isfile(self._get_path(session_id)):
-                with open(self._get_path(session_id), "r") as cin:
+            session_file = self._get_path(session_id)
+            if session_file and os.path.isfile(session_file):
+                with open(session_file, "r") as cin:
                     session = pickle.load(cin)
             else:
                 session = {}
@@ -73,4 +74,7 @@ class FileSessionStorage(SessionStorage):
                 raise SessionInitializeError
 
     def _get_path(self, session_id):
+        if not isinstance(session_id, basestring):
+            log.error("Session id is not string but %r", session_id)
+            return None
         return os.path.join(self._session_dir, session_id)
